Add page for detailed view extra
