+++ pages/detailed-character/[characterId].js
@@ -1,0 +1,25 @@
+import { Title } from '../../components/Header';
+import styled from 'styled-components';
+import DetailedCharacterCard from '../../components/DetailedCharacter';
+import { useRouter } from 'next/router';
+
+export default function DetailedCharacter() {
+  const router = useRouter();
+  const { characterId } = router.query;
+
+  return (
+    <Main>
+      <Title>
+        <img src={'/LogoDragon.svg'} width="35px" />
+        My Characters
+      </Title>
+      <DetailedCharacterCard characterId={characterId} />
+    </Main>
+  );
+}
+
+const Main = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
